SSVPREFAC/DL_Project_refactor/03_code/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
open-clip-torch>=2.24.0
timm>=0.9.12
PyYAML>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0
opencv-python>=4.8.0
matplotlib>=3.8.0
tqdm>=4.66.0
Pillow>=10.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
ftfy>=6.1.1
